+++ orix/quaternion/_conversions.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright 2018-2023 the orix developers
+# Copyright 2018-2024 the orix developers
 #
 # This file is part of orix.
 #
@@ -554,7 +554,7 @@ def ax2qu(axes: np.ndarray, angles: np.ndarray) -> np.ndarray:
 
     if axes.shape[-1] != 3:
         raise ValueError("Final dimension of axes array must be 3.")
-    if angles.shape[-1] != 1 or angles.shape == (1,):
+    if angles.shape[-1] != 1 or angles.shape == (1,) or axes.shape[:-1] == angles.shape:
         angles = angles.reshape(angles.shape + (1,))
 
     axes_shape = axes.shape[:-1]