+++ orix/quaternion/symmetry.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright 2018-2023 the orix developers
+# Copyright 2018-2024 the orix developers
 #
 # This file is part of orix.
 #
@@ -417,7 +417,7 @@ class Symmetry(Rotation):
     def get_highest_order_axis(self) -> Tuple[Vector3d, np.ndarray]:
         axis_orders = self.get_axis_orders()
         if len(axis_orders) == 0:
-            return Vector3d.zvector(), np.infty
+            return Vector3d.zvector(), np.inf
         highest_order = max(axis_orders.values())
         axes = Vector3d.stack(
             [ao for ao in axis_orders if axis_orders[ao] == highest_order]