+++ orix/tests/sampling/test_sampling.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright 2018-2023 the orix developers
+# Copyright 2018-2024 the orix developers
 #
 # This file is part of orix.
 #
@@ -20,8 +20,13 @@ import numpy as np
 import pytest
 
 from orix.quaternion import Rotation
-from orix.quaternion.symmetry import C2, C6, D6, get_point_group
-from orix.sampling import get_sample_fundamental, get_sample_local, uniform_SO3_sample
+from orix.quaternion.symmetry import C2, C4, C6, D6, Oh, get_point_group
+from orix.sampling import (
+    get_sample_fundamental,
+    get_sample_local,
+    get_sample_reduced_fundamental,
+    uniform_SO3_sample,
+)
 from orix.sampling.SO3_sampling import _resolution_to_num_steps
 from orix.sampling._polyhedral_sampling import (
     _get_angles_between_nn_gridpoints,
@@ -29,6 +34,7 @@ from orix.sampling._polyhedral_sampling import (
     _get_max_grid_angle,
     _get_start_and_end_index,
 )
+from orix.vector import Vector3d
 
 
 @pytest.fixture(scope="session")
@@ -175,3 +181,36 @@ class TestSampleFundamental:
         C2_sample = get_sample_fundamental(4, space_group=3, method="haar_euler")
         ratio = C2_sample.size / C6_sample.size
         assert np.isclose(ratio, 3, atol=0.2)
+
+    def test_get_sample_reduced_fundamental(self):
+        vz = Vector3d.zvector()
+
+        R_C1 = get_sample_reduced_fundamental(resolution=4)
+        v_C1 = R_C1 * vz
+        assert np.allclose(v_C1.mean().data, [0, 0, 0])
+
+        R_C4 = get_sample_reduced_fundamental(resolution=4, point_group=C4)
+        v_C4 = R_C4 * vz
+        assert np.all(v_C4 <= C4.fundamental_sector)
+
+        R_C6 = get_sample_reduced_fundamental(resolution=4, point_group=C6)
+        v_C6 = R_C6 * vz
+        assert np.all(v_C6 <= C6.fundamental_sector)
+
+        R_Oh = get_sample_reduced_fundamental(point_group=Oh)
+        v_Oh = R_Oh * vz
+        assert np.all(v_Oh <= Oh.fundamental_sector)
+
+        # Some rotations have a phi1 Euler angle of multiples of pi,
+        # presumably due to rounding errors
+        phi1_C1 = R_C1.to_euler()[:, 0].round(7)
+        assert np.allclose(np.unique(phi1_C1), [0, 2 * np.pi], atol=1e-7)
+        phi1_C4 = R_C4.to_euler()[:, 0].round(7)
+        assert np.allclose(np.unique(phi1_C4), [0, np.pi / 2], atol=1e-7)
+        phi1_C6 = R_C6.to_euler()[:, 0].round(7)
+        assert np.allclose(np.unique(phi1_C6), [0, np.pi / 2], atol=1e-7)
+        phi1_Oh = R_Oh.to_euler()[:, 0].round(7)
+        assert np.allclose(np.unique(phi1_Oh), [0, np.pi / 2], atol=1e-7)
+
+        R_Oh2 = get_sample_reduced_fundamental(point_group=Oh, method="icosahedral")
+        assert R_Oh.size > R_Oh2.size