+++ CHANGELOG.rst
@@ -11,6 +11,41 @@ Unreleased
 
 Added
 -----
+- We can now read 2D crystal maps from Channel Text Files (CTFs) using ``io.load()``.
+
+Changed
+-------
+- Phase names in crystal maps read from .ang files with ``io.load()`` now prefer to use
+  the abbreviated "Formula" instead of "MaterialName" in the file header.
+
+Removed
+-------
+
+Deprecated
+----------
+- ``loadang()`` and ``loadctf()`` are deprecated and will be removed in the next minor
+  release. Please use ``io.load()`` instead.
+
+Fixed
+-----
+
+
+2024-04-21 - version 0.12.1
+===========================
+
+Fixed
+-----
+- ``ax2qu`` and ``Quaternion.from_axes_angles()`` would raise if the input arrays were
+  broadcastable but the final dimension was ``1``. This has been fixed.
+- ``Phase.from_cif()`` now correctly adjusts atom positions when forcing
+  ``Phase.structure.lattice.base`` to use the crystal axes alignment ``e1 || a``,
+  ``e3 || c*``. This bug was introduced in 0.12.0.
+
+2024-04-13 - version 0.12.0
+===========================
+
+Added
+-----
 - ``Vector3d.from_path_ends()`` class method to get vectors between two vectors.
 - Convenience function ``plot.format_labels()`` to get nicely formatted vector labels to
   use when plotting vectors.
@@ -31,6 +66,9 @@ Added
 - The ``random()`` methods of ``Orientation`` and ``Misorientation`` now accept
   ``symmetry``. A ``random()`` method is also added to ``Vector3d`` and ``Miller``, the
   latter accepting a ``phase``.
+- Function ``orix.sampling.get_sample_reduced_fundamental()`` for sampling rotations
+  that rotate the Z-vector (0, 0, 1) onto the fundamental sector of the Laue group of a
+  given ``Symmetry``.
 
 Changed
 -------
@@ -40,19 +78,23 @@ Changed
 - ``random()`` methods no longer accept a list as a valid shape: pass a tuple instead.
 - Increase minimal version of Matplotlib to >= 3.5.
 
+Removed
+-------
+- Support for Python 3.7.
+
 Deprecated
 ----------
 - Creating quaternions from neo-eulerian vectors via ``from_neo_euler()`` is deprecated
   and will be removed in v0.13. Use the existing ``from_axes_angles()`` and the new
   ``from_rodrigues()`` and ``from_homochoric()`` instead.
 
-Removed
--------
-
 Fixed
 -----
 - Transparency of polar stereographic grid lines can now be controlled by Matplotlib's
   ``grid.alpha``, just like the azimuth grid lines.
+- Previously, ``Phase`` did not adjust atom positions when forcing
+  ``Phase.structure.lattice.base`` to use the crystal axes alignment ``e1 || a``,
+  ``e3 || c*``. This is now fixed.
 
 2023-03-14 - version 0.11.1
 ===========================